Spar Polysemy: Separate Misc Utilities (#2069)
* Pull out Http utils

* Pull out some other misc utilities

* make format

* changelog

* CI

* Remove redundant imports

* ci
