+++ changelog.d/5-internal/sem-utils
@@ -1,0 +1,1 @@
+Separate some Spar.Sem utility functions into their own module
