+++ services/spar/src/Spar/Sem/BrigAccess/Http.hs
@@ -24,8 +24,8 @@ import Polysemy.Error (Error)
 import Spar.Error (SparError)
 import qualified Spar.Intra.Brig as Intra
 import Spar.Sem.BrigAccess
-import Spar.Sem.GalleyAccess.Http (RunHttpEnv (..), viaRunHttp)
 import Spar.Sem.Logger (Logger)
+import Spar.Sem.Utils (RunHttpEnv (..), viaRunHttp)
 import qualified System.Logger as TinyLog
 
 brigAccessToHttp ::
